+++ .ci/unit_tests.py
@@ -3,16 +3,10 @@
 # SPDX-License-Identifier: BSD-3-Clause
 from __future__ import absolute_import, division, print_function
 
-import os
 import sys
 
 import ci_lib
 from ci_lib import subprocess
-
-env = os.environ.copy()
-env.update({
-    'PYTHONPATH': '%s%s%s' % (ci_lib.TESTS_DIR, os.sep, env['PYTHONPATH']),
-})
 
 interesting = ci_lib.get_interesting_procs()
 
@@ -25,7 +19,6 @@ with ci_lib.Fold('unit_tests'):
             '--verbose',
         ],
         check=True,
-        env=env,
     )
 
 ci_lib.check_stray_processes(interesting)